Fixes #18276: clean and refactor modules and folders
