+++ rudder-lang/src/command/save.rs
@@ -1,13 +1,15 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// SPDX-FileCopyrightText: 2019-2020 Normation SAS
+
+use super::CommandResult;
 use crate::{
-    error::Result, io::IOContext, logger::*, rudderlang_lib::RudderlangLib, technique::Technique,
-    ActionResult, Format,
+    error::Result, generator::Format, io::IOContext, rudderlang_lib::RudderlangLib,
+    technique::Technique,
 };
-use colored::Colorize;
-use std::{fs, io::Write, path::PathBuf};
 use typed_arena::Arena;
 
 /// Generates RudderLang from JSON technique (constructed around Rudder Technique exported JSON techniques)
-pub fn migrate(ctx: &IOContext) -> Result<Vec<ActionResult>> {
+pub fn save(ctx: &IOContext) -> Result<Vec<CommandResult>> {
     let sources: Arena<String> = Arena::new();
     let json_input = sources.alloc(ctx.input.clone());
     let json_content = sources.alloc(ctx.input_content.clone());
@@ -15,7 +17,7 @@ pub fn migrate(ctx: &IOContext) -> Result<Vec<ActionResult>> {
     let lib = RudderlangLib::new(&ctx.stdlib, &sources)?;
     let technique = Technique::from_json(json_input, json_content, false)?.to_rudderlang(&lib)?;
 
-    Ok(vec![ActionResult::new(
+    Ok(vec![CommandResult::new(
         Format::RudderLang,
         match &ctx.output {
             Some(path) => path.to_str().map(|refstr| refstr.into()), // into a PathBuf
