+++ rudder-lang/src/command/technique_generate.rs
@@ -1,11 +1,15 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// SPDX-FileCopyrightText: 2019-2020 Normation SAS
+
+use super::{compile, CommandResult};
 use crate::{
-    compile::compile, error::Result, io::IOContext, rudderlang_lib::RudderlangLib,
-    technique::Technique, ActionResult, Format,
+    error::Result, generator::Format, io::IOContext, rudderlang_lib::RudderlangLib,
+    technique::Technique,
 };
 use typed_arena::Arena;
 
 /// Takes a JSON technique and generates its RudderLang, DSC< CFEngine equivalent wrapped into a single JSON object
-pub fn technique_generate(ctx: &IOContext) -> Result<Vec<ActionResult>> {
+pub fn technique_generate(ctx: &IOContext) -> Result<Vec<CommandResult>> {
     // push RL version to the final data object
     let sources: Arena<String> = Arena::new();
     let input = sources.alloc(format!("JSON based on {}", ctx.input));
@@ -13,7 +17,7 @@ pub fn technique_generate(ctx: &IOContext) -> Result<Vec<ActionResult>> {
     let lib = RudderlangLib::new(&ctx.stdlib, &sources)?;
     let technique_fmt = Technique::from_json(input, &content, false)?.to_rudderlang(&lib)?;
 
-    let mut wrapped_technique = vec![ActionResult::new(
+    let mut wrapped_technique = vec![CommandResult::new(
         Format::RudderLang,
         match &ctx.output {
             Some(path) => path.to_str().map(|refstr| refstr.into()), // into a PathBuf
@@ -29,7 +33,7 @@ pub fn technique_generate(ctx: &IOContext) -> Result<Vec<ActionResult>> {
     // push dsc version to the final data object
     wrapped_technique.extend(compile(&updated_ctx.with_format(Format::DSC), true)?);
 
-    // TODO do not stop at first error: accumulate errors, create each ActionResult (empty) and return all errors
+    // TODO do not stop at first error: accumulate errors, create each CommandResult (empty) and return all errors
     // requires a working log / error system
     Ok(wrapped_technique)
 }
