+++ rudder-lang/src/io/logs/levels.rs
@@ -1,0 +1,54 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// SPDX-FileCopyrightText: 2019-2020 Normation SAS
+
+use crate::error::*;
+use serde::{Deserialize, Serialize};
+use std::{fmt, str::FromStr};
+
+#[derive(Serialize, Deserialize, Copy, Clone, Debug, Eq, PartialEq, Ord, PartialOrd)]
+pub enum LogLevel {
+    Error,
+    Warn,
+    Info,
+    Debug,
+    Trace,
+}
+impl LogLevel {
+    pub fn includes(&self, other: LogLevel) -> bool {
+        self >= &other
+    }
+}
+impl Default for LogLevel {
+    fn default() -> Self {
+        Self::Warn
+    }
+}
+impl fmt::Display for LogLevel {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                Self::Error => "ERROR",
+                Self::Warn => "WARN",
+                Self::Info => "INFO",
+                Self::Debug => "DEBUG",
+                Self::Trace => "TRACE",
+            }
+        )
+    }
+}
+impl FromStr for LogLevel {
+    type Err = Error;
+
+    fn from_str(input: &str) -> Result<Self> {
+        match input.to_lowercase().as_ref() {
+            "error" => Ok(Self::Error),
+            "warn" => Ok(Self::Warn),
+            "info" => Ok(Self::Info),
+            "debug" => Ok(Self::Debug),
+            "trace" => Ok(Self::Trace),
+            _ => Err(Error::new(format!("Could not parse log level '{}'", input))),
+        }
+    }
+}
