+++ rudder-lang/src/io/output.rs
@@ -1,104 +1,22 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // SPDX-FileCopyrightText: 2019-2020 Normation SAS
 
-use crate::{error::Result, logger::*, Action, ActionResult};
+mod backtrace;
+
+pub use self::backtrace::Backtrace;
+
+use super::logs::*;
+use crate::{command::*, error::Result};
 use colored::Colorize;
 use lazy_static::lazy_static;
 use regex::Regex;
 use serde::Serialize;
-use std::fmt::Display;
 use std::{
-    fmt,
     fs::File,
     io::Write,
     panic,
     time::{SystemTime, UNIX_EPOCH},
 };
-
-#[derive(Clone)]
-pub struct Backtrace(backtrace::Backtrace);
-
-impl Backtrace {
-    pub fn new() -> Self {
-        Self(backtrace::Backtrace::new())
-    }
-
-    /// this funtion can be invoked from anywhere in the code to get a proper backtrace up to the creation of the programs thread
-    /// This could be done for debug purposes or placed at strategic places like panic calls
-    fn format_symbol(index: usize, sym: &backtrace::BacktraceSymbol) -> Option<String> {
-        lazy_static! {
-            // if starts with rudderc + remove ending addr that is not helpful as is + handle aliases
-            static ref SYMBOL_PATH: Regex = Regex::new(r"^<?(?P<path>rudderc(::[{}\d\w]+)+)( as .*>(?P<endingpath>::[{}\d\w]+)+)?(::[\da-z]+)$").unwrap();
-        }
-
-        Self::get_symbol_name(sym).and_then(|str_name| {
-            SYMBOL_PATH
-                .captures(&str_name)
-                .and_then(|caps| match (caps.name("path"), caps.name("endingpath")) {
-                    (None, _) => None,
-                    (Some(start), Some(end)) => [start.as_str(), end.as_str()].concat().into(),
-                    (Some(start), None) => start.as_str().to_owned().into(),
-                })
-                .and_then(|fmt_name| {
-                    // do not put output related calls in the backtrace since it always ultimately calls panic_hook and print_backtrace
-                    if fmt_name.starts_with("rudderc::output")
-                        || fmt_name.starts_with("rudderc::error::Error::new")
-                    {
-                        return None;
-                    }
-                    Some(format!(
-                        "  {offset}{name} at '{filename}:{line}'",
-                        offset = " ".repeat(index * 2),
-                        name = fmt_name,
-                        filename = Self::get_symbol_filename(sym),
-                        line = Self::get_symbol_line(sym)
-                    ))
-                })
-        })
-    }
-
-    fn get_symbol_name(sym: &backtrace::BacktraceSymbol) -> Option<String> {
-        sym.name().and_then(|name| format!("{:?}", name).into())
-    }
-
-    fn get_symbol_line(sym: &backtrace::BacktraceSymbol) -> String {
-        sym.lineno()
-            .and_then(|n| n.to_string().into())
-            .unwrap_or("undefined".to_owned())
-    }
-
-    fn get_symbol_filename(sym: &backtrace::BacktraceSymbol) -> String {
-        sym.filename()
-            .and_then(|path| path.to_str())
-            .unwrap_or("undefined")
-            .to_owned()
-    }
-
-    // TODO backtrace as vec, to print it into a backtrace field in the json format
-}
-
-/// Display backtrace to the final user
-impl Display for Backtrace {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let stringified_trace = self
-            .0
-            .frames()
-            .iter()
-            .filter_map(|frame| {
-                frame
-                    .symbols()
-                    .into_iter()
-                    .enumerate()
-                    .map(|(index, sym)| Self::format_symbol(index, sym))
-                    .collect::<Option<Vec<String>>>()
-            })
-            .flatten()
-            .collect::<Vec<String>>()
-            .join("\n");
-
-        write!(f, "\nTrace:\n{}", stringified_trace)
-    }
-}
 
 // NOTE that OutputFmtPanic is hardcoded, not serializable structure to pass to it
 // struct OutputFmtPanic {
@@ -108,12 +26,12 @@ impl Display for Backtrace {
 /// Note that `source` field is either a file or `STDIN`. It can even be empty in some error cases
 #[derive(Serialize)]
 struct OutputFmtOk {
-    action: String,
+    command: String,
     source: String, // source file path or STDIN
     time: String,
     status: String, // either "success" or "failure"
     logs: Logs,
-    data: Vec<ActionResult>,
+    data: Vec<CommandResult>,
     errors: Vec<String>,
 }
 
@@ -125,10 +43,10 @@ pub enum LogOutput {
 }
 impl LogOutput {
     /// Initialize the output generator
-    pub fn init(self, action: Action, log_level: LogLevel, is_backtraced: bool) {
+    pub fn init(self, command: Command, log_level: LogLevel, is_backtraced: bool) {
         set_global_log_level(log_level);
         // Content called if panic! is encountered to close logger brackets and print error
-        self.set_panic_hook(action, is_backtraced);
+        self.set_panic_hook(command, is_backtraced);
         // enable rust backtrace if backtrace requested
         if is_backtraced {
             std::env::set_var("RUDDERC_BACKTRACE", "1");
@@ -162,7 +80,7 @@ impl LogOutput {
 
     /// panic default format takeover to print either proper json format output
     /// or rudder-lang own error logging format
-    fn set_panic_hook(self, action: Action, is_backtraced: bool) {
+    fn set_panic_hook(self, command: Command, is_backtraced: bool) {
         panic::set_hook(Box::new(move |panic_info| {
             let e_message = match panic_info.payload().downcast_ref::<&str>() {
                 Some(msg) => msg.to_string(), // PANIC!
@@ -188,7 +106,7 @@ impl LogOutput {
             };
 
             self.print(
-                action,
+                command,
                 String::new(),
                 Err(crate::error::Error::new(message)),
             );
@@ -198,7 +116,7 @@ impl LogOutput {
         }));
     }
 
-    pub fn print(self, action: Action, source: String, result: Result<Vec<ActionResult>>) {
+    pub fn print(self, command: Command, source: String, result: Result<Vec<CommandResult>>) {
         let (is_success, mut data, errors) = match result {
             Ok(data) => (true, data, Vec::new()),
             Err(e) => (false, Vec::new(), e.clean_format_list()),
@@ -217,7 +135,7 @@ impl LogOutput {
         };
 
         // remove data if it has been written to a file to avoid duplicate content
-        self.print_to_file(&mut data, action);
+        self.print_to_file(&mut data, command);
 
         if !is_success {
             let output = format!(
@@ -237,7 +155,7 @@ impl LogOutput {
         if self == LogOutput::JSON {
             let status = if is_success { "success" } else { "failure" };
             let output = OutputFmtOk {
-                action: format!("{}", action),
+                command: format!("{}", command),
                 time,
                 status: status.to_owned(),
                 source,
@@ -253,9 +171,9 @@ impl LogOutput {
     }
 
     // print content into a file and if successfully written, remove it from payload
-    fn print_to_file(&self, files: &mut Vec<ActionResult>, action: Action) {
+    fn print_to_file(&self, files: &mut Vec<CommandResult>, command: Command) {
         if self == &LogOutput::None {
-            if action == Action::GenerateTechnique {
+            if command == Command::GenerateTechnique {
                 println!(
                     "{}",
                     serde_json::to_string_pretty(&files)
