+++ rudder-lang/tools/rudderc-dev.conf
@@ -10,7 +10,7 @@ dsc_methods="repos/dsc/packaging/Files/share/initial-policy/ncf/30_generic_metho
 input="tests/techniques/"
 output="tests/techniques/"
 
-[migrate]
+[save]
 input="tests/techniques/"
 output="tests/techniques/"
 
@@ -26,4 +26,3 @@ output="tests/techniques/"
 cfengine="/opt/rudder/bin/cf-promises"
 ncf_tools="repos/ncf/tools/"
 py_modules="tools/"
-
