+++ rudder-lang/tools/rudderc-prod.conf
@@ -9,7 +9,7 @@ stdlib ="/opt/rudder/share/rudder-lang/lib/"
 input="/var/rudder/configuration-repository/techniques/ncf_techniques/"
 output="/tmp/rudderc/tester/"
 
-[migrate]
+[save]
 input="/tmp/rudderc/tester/"
 output="/tmp/rudderc/tester/"
